requirements.txt
discord.py
dispander
